Enum nterm::Event

pub enum Event {
    Key(Key),
    Character(char),
    Resize(u32, u32),
    MouseDown(u32, u32, MouseButton),
    MouseUp(u32, u32, MouseButton),
}

All events that can occur.

Variants

Key

When a non-character key is pressed.

Character

When a character is key is pressed.

Resize

When the terminal window is resized. Passed the new width and height of the terminal.

MouseDown

When a mouse button is pressed.

MouseUp

When a mouse button is released after being pressed.