Struct nterm::Terminal

pub struct Terminal;

The terminal itself.

Methods

impl Terminal

fn new() -> Terminal

Create a new instance of the termial window. Usually is only called once upon program start to initialise ncurses.

fn cursor_state(&self, state: Cursor)

Change the cursor state.

fn has_mouse_support(&self) -> bool

Returns true if the mouse is supported on this terminal.

fn poll_event(&self) -> Event

Waits for an event to occur, returning the event when one does.